requirements.txt
fastapi
uvicorn
pydantic
pymongo
dnspython
python-jose
passlib[bcrypt]
slowapi


